+++ lib/dataloaders/index.ts
@@ -1,1 +1,3 @@
 export * from './references-many.loader';
+export * from './has-many.loader';
+export * from './has-one.loader';