+++ lib/property.ts
@@ -99,8 +99,82 @@ export function ReferencesMany(
   };
 }
 
+export type HasOneConfig = {
+  typeFunction: () => any;
+  options: {
+    to: string;
+  };
+};
+export function HasOne(typeFunction: HasOneConfig['typeFunction'], options: HasOneConfig['options']) {
+  return (target: object, propertyKey: string | symbol) => {
+    ExcludeOnDatabase()(target, propertyKey);
+    Metadata.for(target).with(propertyKey).set<HasOneConfig>(MetaKey.HasOneType, { typeFunction, options });
+  };
+}
+
+export type HasManyConfig = {
+  typeFunction: () => any;
+  options: {
+    from: string;
+  };
+};
+export function HasMany(typeFunction: HasManyConfig['typeFunction'], options: HasManyConfig['options']) {
+  return (target: object, propertyKey: string | symbol) => {
+    ExcludeOnDatabase()(target, propertyKey);
+    Metadata.for(target).with(propertyKey).set<HasManyConfig>(MetaKey.HasManyType, { typeFunction, options });
+  };
+}
+
 export function ExcludeOnDatabase() {
   return (target: object, propertyKey: string | symbol) => {
     Metadata.for(target).with(propertyKey).set(MetaKey.ExcludeOnDatabase, true);
+  };
+}
+
+export type FilterOperator =
+  | 'eq'
+  | 'in'
+  | 'notEq'
+  | 'notIn'
+  | 'contains'
+  | 'notContains'
+  | 'gt'
+  | 'gte'
+  | 'lt'
+  | 'lte'
+  | 'regex'
+  | 'notRegex'
+  | 'all'
+  | 'exists';
+
+export const allOperators: FilterOperator[] = [
+  'eq',
+  'in',
+  'notEq',
+  'notIn',
+  'contains',
+  'notContains',
+  'gt',
+  'gte',
+  'lt',
+  'lte',
+  'regex',
+  'notRegex',
+  'all',
+  'exists',
+];
+
+export function Filterable(typeFn: () => any, input: { operators: FilterOperator[] }) {
+  return (target: object, propertyKey: string | symbol) => {
+    Metadata.for(target).with(propertyKey).set(MetaKey.Filterable, {
+      typeFn,
+      input,
+    });
+  };
+}
+
+export function Sortable() {
+  return (target: object, propertyKey: string | symbol) => {
+    Metadata.for(target).with(propertyKey).set(MetaKey.Sortable, true);
   };
 }