+++ src/models/product.ts
@@ -11,7 +11,14 @@ import {
   Thunk,
   OutputType,
   CreateInputType,
+  Filterable,
+  Sortable,
+  HasMany,
+  ExcludeOnDatabase,
+  HasOne,
 } from '../../lib';
+import { MaxLength } from 'class-validator';
+import { Variant } from './variant';
 
 @Definition({ allowedApis: '*' })
 export class Tag {
@@ -19,8 +26,26 @@ export class Tag {
   id: string;
 
   @Prop({ unique: true })
+  @Thunk(Field(() => graphql.GraphQLString))
+  @Thunk(MaxLength(100), { scopes: 'input' })
+  @Thunk(Transform(({ value }) => value.trim()), { scopes: 'input' })
+  name: string;
+}
+
+@Definition({ allowedApis: '*' })
+export class Image {
+  @Property(() => graphql.GraphQLID)
+  id: string;
+
+  @Prop()
   @Property(() => graphql.GraphQLString)
   name: string;
+
+  @Prop({
+    type: ObjectId,
+  })
+  @Thunk(Field(() => graphql.GraphQLID), { scopes: 'output' })
+  productId: string;
 }
 
 @Definition({ allowedApis: '*' })
@@ -29,6 +54,8 @@ export class Product {
   id: string;
 
   @Property(() => graphql.GraphQLString)
+  @Filterable(() => graphql.GraphQLString, { operators: ['eq'] })
+  @Sortable()
   @Prop()
   name: string;
 
@@ -48,4 +75,16 @@ export class Product {
   @Thunk(Field(() => [OutputType(Tag)]), { scopes: 'output' })
   @Thunk(Field(() => [CreateInputType(Tag)], { nullable: true }), { scopes: 'create' })
   tags: Tag[];
+
+  @HasMany(() => Variant, { from: 'productId' })
+  @Thunk(Field(() => [OutputType(Variant)]), { scopes: 'output' })
+  @Thunk(Field(() => [CreateInputType(Variant)], { nullable: true }), { scopes: 'create' })
+  @ExcludeOnDatabase()
+  variants: Variant[];
+
+  @HasOne(() => Image, { to: 'productId' })
+  @Thunk(Field(() => OutputType(Image), { nullable: true }), { scopes: 'output' })
+  @Thunk(Field(() => CreateInputType(Image), { nullable: true }), { scopes: 'create' })
+  @ExcludeOnDatabase()
+  image: Image;
 }